Chore: make DoubleEndedBuffer unconstrained
Update `DoubleEndedBuffer` to support an unconstrained buffer size and make this the default. To make the buffer unconstrained, set `highWaterMark` to `undefined` in the constructor options (the default).

When `highWaterMark` is set to a value, calls to `write` will block if the internal buffer has a length greater than the `highWaterMark`. If no value is provided, then calls to `write` will never block.